backend/requirements.txt
Flask
python-docx
openpyxl
pandas
flask-cors